+++ src/grocker/resources/docker/runner-image/provision.sh
@@ -10,8 +10,7 @@ function setup_venv() {  # venv runtime *dependencies
     shift 2
     local release="$*"
     local constraint_arg=$([ -f ${WORKING_DIR}/constraints.txt ] && echo "--constraint ${WORKING_DIR}/constraints.txt")
-    local grocker_pypi_ip=$(cat ${WORKING_DIR}/pypi.ip)
-    local wheelhouse=http://${grocker_pypi_ip}/
+    local wheelhouse=http://${GROCKER_PYPI_IP}/
 
     local pip=${venv}/bin/pip
 
@@ -19,7 +18,7 @@ function setup_venv() {  # venv runtime *dependencies
     # Old pip can not deal with constraint file
     ${pip} install --upgrade pip
     ${pip} install --upgrade pip setuptools ${constraint_arg}
-    ${pip} install --find-links=${wheelhouse} --trusted-host=${grocker_pypi_ip} --no-index ${constraint_arg} ${release} --no-compile
+    ${pip} install --find-links=${wheelhouse} --trusted-host=${GROCKER_PYPI_IP} --no-index ${constraint_arg} ${release} --no-compile
 }